replace react-query w/ lighter suspect-react, dont use hooks, return data rather than {data}
