+++ client/package.json
@@ -77,7 +77,6 @@
     "react-flip-move": "^3.0.4",
     "react-focus-lock": "^2.5.0",
     "react-intersection-observer": "^8.30.1",
-    "react-query": "^3.39.0",
     "react-redux": "^7",
     "react-resize-detector": "^6.6.0",
     "react-responsive": "^8.2",
@@ -92,6 +91,7 @@
     "reselect": "^3.0.1",
     "resize-observer": "^1.0.4",
     "string-hash": "^1.1.3",
+    "suspend-react": "^0.0.8",
     "webfontloader": "^1.6.28",
     "whatwg-fetch": "^2.0.4"
   },
