--- client/src/graphql_utils/react_query_client.ts
@@ -1,9 +1,0 @@
-import { QueryClient as ReactQueryClient } from "react-query";
-
-export const reactQueryClient = new ReactQueryClient({
-  defaultOptions: {
-    queries: {
-      suspense: true,
-    },
-  },
-});
